+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.3.0-dev - Current master
 
+- added intel.handlers.Rotating
 - added ignore options to intel.console()
 
 ## v0.2.0 - 2013-10-04